Change project name from `spring-boot-app-template` to `spring-starter`
This commit changes the project name to a more easy to find and remember name that reflects better on the purpose of this project.
