--- buildSrc/src/main/groovy/spring-boot-app-template.publishing-conventions.gradle
@@ -1,26 +1,0 @@
-plugins {
-    id('maven-publish')
-}
-
-publishing {
-    publications {
-        maven(MavenPublication) {
-            groupId = 'com.yonatankarp'
-            artifactId = project.name
-            version = project.version
-
-            from components.java
-        }
-    }
-
-    repositories {
-        maven {
-            name = 'GitHubPackages'
-            url = uri("https://maven.pkg.github.com/sumup/spring-boot-app-template")
-            credentials {
-                username = project.findProperty('gpr.user') ?: System.getenv('GITHUB_ACTOR')
-                password = project.findProperty('gpr.key') ?: System.getenv('GITHUB_TOKEN')
-            }
-        }
-    }
-}
