--- spring-boot-app-template/src/main/kotlin/com/yonatankarp/springbootapptemplate/application/ports/HelloWorldPort.kt
@@ -1,5 +1,0 @@
-package com.yonatankarp.springbootapptemplate.application.ports
-
-interface HelloWorldPort {
-    suspend fun greet(): String
-}
