--- spring-boot-app-template/src/main/kotlin/com/yonatankarp/springbootapptemplate/application/usecases/HelloWorldUseCase.kt
@@ -1,10 +1,0 @@
-package com.yonatankarp.springbootapptemplate.application.usecases
-
-import com.yonatankarp.springbootapptemplate.application.ports.HelloWorldPort
-import com.yonatankarp.springbootapptemplate.domain.valueobject.HelloWorld
-import org.springframework.stereotype.Service
-
-@Service
-class HelloWorldUseCase : HelloWorldPort {
-    override suspend fun greet() = HelloWorld().sayHello()
-}
