+++ spring-starter/src/main/kotlin/com/yonatankarp/spring/starter/application/ports/HelloWorldPort.kt
@@ -1,0 +1,5 @@
+package com.yonatankarp.spring.starter.application.ports
+
+interface HelloWorldPort {
+    suspend fun greet(): String
+}
