+++ spring-starter/src/main/kotlin/com/yonatankarp/spring/starter/domain/valueobject/HelloWorld.kt
@@ -1,0 +1,5 @@
+package com.yonatankarp.spring.starter.domain.valueobject
+
+class HelloWorld {
+    fun sayHello(): String = "Hello, World!"
+}
