+++ spring-starter/src/main/resources/api/demo-api.yaml
