+++ spring-starter/src/main/resources/api/templates/kotlin-spring/apiInterface.mustache
