+++ spring-starter/src/main/resources/db/migration/V1.0.0__init_db.sql
