+++ spring-starter/src/main/resources/logback-stdout-appender.xml
